job-platform-demo-backend/pyproject.toml
[project]
name = "job-platform-demo-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Peter Chen",email = "sanpeter992000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-ninja (>=1.4.1,<2.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"